requirements.txt
django==4.2.3
tensorflow==2.17.0
keras==3.2.0
opencv-python==4.7.0.72
joblib==1.2.0
Pillow==10.0.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.12.2
scikit-learn==1.2.0
IPython==8.14.0
